update content tools from android to desktop

+++ src/main/java/com/mytlogos/enterprisedesktop/tools/ContentTool.java
@@ -12,14 +12,12 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public abstract class ContentTool {
-    final File internalContentDir;
-    final File externalContentDir;
+    final File contentDir;
     final int minMBSpaceAvailable = 150;
 
 
-    ContentTool(File internalContentDir, File externalContentDir) {
-        this.internalContentDir = internalContentDir;
-        this.externalContentDir = externalContentDir;
+    ContentTool(File contentDir) {
+        this.contentDir = contentDir;
     }
 
     /**
@@ -36,23 +34,25 @@ public abstract class ContentTool {
     public List<String> getMediaPaths() {
         List<String> books = new ArrayList<>();
 
-        if (externalContentDir != null) {
-            books.addAll(this.getMediaPaths(externalContentDir));
-        }
-        if (internalContentDir != null) {
-            books.addAll(this.getMediaPaths(internalContentDir));
+        if (this.contentDir != null) {
+            books.addAll(this.getMediaPaths(contentDir));
         }
         return books;
     }
 
     private List<String> getMediaPaths(File dir) {
-        List<String> imageMedia = new ArrayList<>();
-        for (File file : dir.listFiles()) {
+        List<String> mediaPaths = new ArrayList<>();
+        final File[] files = dir.listFiles();
+
+        if (files == null) {
+            return mediaPaths;
+        }
+        for (File file : files) {
             if (isContentMedium(file)) {
-                imageMedia.add(file.getAbsolutePath());
+                mediaPaths.add(file.getAbsolutePath());
             }
         }
-        return imageMedia;
+        return mediaPaths;
     }
 
     abstract boolean isContentMedium(File file);
@@ -60,18 +60,11 @@ public abstract class ContentTool {
     public abstract int getMedium();
 
     public void removeMediaEpisodes(int mediumId, Set<Integer> episodeIds) {
-        if (externalContentDir != null) {
-            String externalFile = getItemPath(mediumId, externalContentDir);
+        if (this.contentDir != null) {
+            String itemPath = getItemPath(mediumId, this.contentDir);
 
-            if (externalFile != null) {
-                this.removeMediaEpisodes(episodeIds, externalFile);
-            }
-        }
-        if (internalContentDir != null) {
-            String internalFile = getItemPath(mediumId, internalContentDir);
-
-            if (internalFile != null) {
-                this.removeMediaEpisodes(episodeIds, internalFile);
+            if (itemPath != null) {
+                this.removeMediaEpisodes(episodeIds, itemPath);
             }
         }
     }
@@ -85,12 +78,8 @@ public abstract class ContentTool {
     public String getItemPath(int mediumId) {
         String bookZipFile = null;
 
-        if (externalContentDir != null) {
-            bookZipFile = getItemPath(mediumId, externalContentDir);
-        }
-
-        if (bookZipFile == null && internalContentDir != null) {
-            bookZipFile = getItemPath(mediumId, internalContentDir);
+        if (this.contentDir != null) {
+            bookZipFile = getItemPath(mediumId, this.contentDir);
         }
 
         if (bookZipFile == null) {
@@ -101,36 +90,10 @@ public abstract class ContentTool {
 
     public abstract void saveContent(Collection<ClientDownloadedEpisode> episode, int mediumId) throws IOException;
 
-    public void mergeIfNecessary() {
-        if (!this.isSupported()) {
-            return;
-        }
-        if (this.writeInternal() && !this.writeExternal()) {
-            this.mergeExternalAndInternalMedia(false);
-        } else if (!this.writeInternal() && this.writeExternal()) {
-            this.mergeExternalAndInternalMedia(true);
-        }
-    }
-
     public abstract boolean isSupported();
 
-    synchronized public void mergeExternalAndInternalMedia(boolean toExternal) {
-        Map<Integer, File> internalContainers = this.getItemContainers(false);
-        Map<Integer, File> externalContainers = this.getItemContainers(true);
-
-        Map<Integer, File> sourceContainers = toExternal ? internalContainers : externalContainers;
-        Map<Integer, File> toContainers = toExternal ? externalContainers : internalContainers;
-
-        File toParent = toExternal ? externalContentDir : internalContentDir;
-
-        for (Map.Entry<Integer, File> entry : sourceContainers.entrySet()) {
-            File file = toContainers.get(entry.getKey());
-            this.mergeExternalAndInternalMedium(toExternal, entry.getValue(), file, toParent, entry.getKey());
-        }
-    }
-
-    public Map<Integer, File> getItemContainers(boolean externalSpace) {
-        File file = externalSpace ? externalContentDir : internalContentDir;
+    public Map<Integer, File> getItemContainers() {
+        File file = this.contentDir;
 
         if (file == null) {
             return new HashMap<>();
@@ -140,6 +103,10 @@ public abstract class ContentTool {
 
         Map<Integer, File> mediumIdFileMap = new HashMap<>();
 
+        if (files == null) {
+            System.err.println("could not list files for: " + file);
+            return mediumIdFileMap;
+        }
         for (File bookFile : files) {
             Matcher matcher = pattern.matcher(bookFile.getName());
 
@@ -160,22 +127,13 @@ public abstract class ContentTool {
         return mediumIdFileMap;
     }
 
-    abstract void mergeExternalAndInternalMedium(boolean toExternal, File source, File goal, File toParent, Integer mediumId);
-
     abstract Pattern getMediumContainerPattern();
 
     abstract int getMediumContainerPatternGroup();
 
     public void removeMedia(int id) {
-        if (externalContentDir != null) {
-            String externalFile = getItemPath(id, externalContentDir);
-
-            if (externalFile != null && !new File(externalFile).delete()) {
-                System.err.println("could not delete file: " + externalFile);
-            }
-        }
-        if (internalContentDir != null) {
-            String internalFile = getItemPath(id, internalContentDir);
+        if (this.contentDir != null) {
+            String internalFile = getItemPath(id, this.contentDir);
 
             if (internalFile != null && !new File(internalFile).delete()) {
                 System.err.println("could not delete file: " + internalFile);
@@ -184,15 +142,12 @@ public abstract class ContentTool {
     }
 
     public void removeAll() {
-        for (File file : this.internalContentDir.listFiles()) {
-            if (file.isDirectory()) {
-                this.deleteDir(file);
-            }
-            if (file.exists() && !file.delete()) {
-                System.err.println("could not delete file: " + file.getAbsolutePath());
-            }
+        final File[] files = this.contentDir.listFiles();
+        if (files == null) {
+            System.err.println("could not list files of: " + this.contentDir);
+            return;
         }
-        for (File file : this.externalContentDir.listFiles()) {
+        for (File file : files) {
             if (file.isDirectory()) {
                 this.deleteDir(file);
             }
@@ -211,39 +166,28 @@ public abstract class ContentTool {
     public abstract double getAverageEpisodeSize(int mediumId);
 
     boolean writeable() {
-        return this.writeExternal() || this.writeInternal();
-    }
-
-    boolean writeExternal() {
-        return externalContentDir != null && externalContentDir.getUsableSpace() >= minByteSpaceAvailable();
-    }
-
-    boolean writeInternal() {
-        return internalContentDir != null && internalContentDir.getUsableSpace() >= minByteSpaceAvailable();
+        return contentDir != null && contentDir.getUsableSpace() >= minByteSpaceAvailable();
     }
 
     private long minByteSpaceAvailable() {
         return minMBSpaceAvailable * 1024 * 1024;
     }
 
-    boolean writeable(long toWriteBytes) {
-        return writeExternal(toWriteBytes) || writeInternal(toWriteBytes);
+    boolean notWriteable(long toWriteBytes) {
+        return this.contentDir != null && (this.contentDir.getUsableSpace() - toWriteBytes) >= minByteSpaceAvailable();
     }
 
-    boolean writeExternal(long toWriteBytes) {
-        return externalContentDir != null && (externalContentDir.getUsableSpace() - toWriteBytes) >= minByteSpaceAvailable();
-    }
-
-    boolean writeInternal(long toWriteBytes) {
-        return internalContentDir != null && (internalContentDir.getUsableSpace() - toWriteBytes) >= minByteSpaceAvailable();
-    }
-
-    boolean writeable(File file, long toWriteBytes) {
-        return file != null && (file.getUsableSpace() - toWriteBytes) >= minByteSpaceAvailable();
+    boolean notWriteable(File file, long toWriteBytes) {
+        return file == null || (file.getUsableSpace() - toWriteBytes) < minByteSpaceAvailable();
     }
 
     private void deleteDir(File file) {
-        for (File content : file.listFiles()) {
+        final File[] files = file.listFiles();
+        if (files == null) {
+            System.err.println("could not list files of: " + file);
+            return;
+        }
+        for (File content : files) {
             if (content.isDirectory()) {
                 this.deleteDir(content);
             }
